+++ foodWasteReductionPlatform/src/main/java/functions/Logger.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package dataAccess;
+package functions;
 
 import java.io.BufferedWriter;
 import java.io.FileWriter;
